Positional Control Matlab File - 3 Nov 2023 (1st Official)
A few input parameters and states: 
x parameter: Runs as a float and signifies 12 states in this file

Input parameters inserted: 
1) x  = all 12 states (12 parameters to be recorded)
2) xyz_desired = desired positions of x y and z 
3) param - aka parameters to be inserted for constants. 

Output parameters :
1) s_x 
2) s_y
3) s_z 
The above 3 parameters are put into the attitude controller next to determine the desired phi max, theta max and also the psi max for the drone. 

4)thrustz - this thrustz parameter is using (3) to do equation (We have neglected desired cos phi and desired cos theta in this equation as both are approx = 1) - Refer to Paper on Hossan paper - Equation 25.
Thrust can also be on the positive end - might need to check for the rationale on why using absolute and sign is there inside the matlab code.

5) sum_error_x 
6) sum_error_y
7) sum_error_z

The 3 above parameters, (5),(6) and (7) revises the error again and error for x y and z to be inputted into s_x,s_y,s_z respectively.

You can call s_x,s_y,s_z indirectly as either the force of x,y,z or the virtual control for x,y,z axis. (need to double check and confirm this again with Dr Huang for this)